+++ examples/advanced/src/gen/models/ts/Order.ts
@@ -34,11 +34,15 @@ export type OrderHttpStatusEnumKey = (typeof orderHttpStatusEnum)[keyof typeof o
  */
 export type Order = {
   /**
+   * @description
+   * Format: `int64`
    * @example 10
    * @type integer | undefined
    */
   id?: number
   /**
+   * @description
+   * Format: `int64`
    * @example 198772
    * @type integer | undefined
    */
@@ -59,6 +63,8 @@ export type Order = {
     type: string
   }
   /**
+   * @description
+   * Format: `int32`
    * @example 7
    * @type integer | undefined
    */
@@ -74,6 +80,8 @@ export type Order = {
    */
   type?: string
   /**
+   * @description
+   * Format: `date-time`
    * @type string | undefined
    */
   shipDate?: string