+++ examples/advanced/src/gen/models/ts/petController/UploadFile.ts
@@ -2,6 +2,8 @@ import type { ApiResponse } from '../ApiResponse.ts'
 
 /**
  * @description ID of pet to update
+ *
+ * Format: `int64`
  * @type integer
  */
 export type UploadFilePathPetId = number
@@ -18,6 +20,8 @@ export type UploadFileQueryAdditionalMetadata = string | undefined
 export type UploadFileStatus200 = ApiResponse
 
 /**
+ * @description
+ * Format: `binary`
  * @type string
  */
 export type UploadFileData = Blob