+++ examples/cypress/src/gen-v4/models.ts
@@ -24,21 +24,29 @@ export type OrderHttpStatusEnumKey = (typeof orderHttpStatusEnum)[keyof typeof o
  */
 export type Order = {
   /**
+   * @description
+   * Format: `int64`
    * @example 10
    * @type integer | undefined
    */
   id?: bigint
   /**
+   * @description
+   * Format: `int64`
    * @example 198772
    * @type integer | undefined
    */
   petId?: bigint
   /**
+   * @description
+   * Format: `int32`
    * @example 7
    * @type integer | undefined
    */
   quantity?: number
   /**
+   * @description
+   * Format: `date-time`
    * @type string | undefined
    */
   shipDate?: string
@@ -91,6 +99,8 @@ export type Address = {
  */
 export type Customer = {
   /**
+   * @description
+   * Format: `int64`
    * @example 100000
    * @type integer | undefined
    */
@@ -111,6 +121,8 @@ export type Customer = {
  */
 export type Category = {
   /**
+   * @description
+   * Format: `int64`
    * @example 1
    * @type integer | undefined
    */
@@ -127,6 +139,8 @@ export type Category = {
  */
 export type User = {
   /**
+   * @description
+   * Format: `int64`
    * @example 10
    * @type integer | undefined
    */
@@ -163,6 +177,8 @@ export type User = {
   phone?: string
   /**
    * @description User Status
+   *
+   * Format: `int32`
    * @example 1
    * @type integer | undefined
    */
@@ -174,6 +190,8 @@ export type User = {
  */
 export type Tag = {
   /**
+   * @description
+   * Format: `int64`
    * @type integer | undefined
    */
   id?: bigint
@@ -196,6 +214,8 @@ export type PetStatusEnumKey = (typeof petStatusEnum)[keyof typeof petStatusEnum
  */
 export type Pet = {
   /**
+   * @description
+   * Format: `int64`
    * @example 10
    * @type integer | undefined
    */
@@ -237,6 +257,8 @@ export type AddPetRequestStatusEnumKey = (typeof addPetRequestStatusEnum)[keyof 
  */
 export type AddPetRequest = {
   /**
+   * @description
+   * Format: `int64`
    * @example 10
    * @type integer | undefined
    */
@@ -270,6 +292,8 @@ export type AddPetRequest = {
  */
 export type ApiResponse = {
   /**
+   * @description
+   * Format: `int32`
    * @type integer | undefined
    */
   code?: number
@@ -288,6 +312,8 @@ export type ApiResponse = {
  */
 export type PetNotFound = {
   /**
+   * @description
+   * Format: `int32`
    * @type integer | undefined
    */
   code?: number
@@ -381,6 +407,8 @@ export type AddPetStatus200 = Pet
  */
 export type AddPetStatus405 = {
   /**
+   * @description
+   * Format: `int32`
    * @type integer | undefined
    */
   code?: number
@@ -589,6 +617,8 @@ export type FindPetsByTagsResponse = FindPetsByTagsStatus200 | FindPetsByTagsSta
 
 /**
  * @description ID of pet to return
+ *
+ * Format: `int64`
  * @type integer
  */
 export type GetPetByIdPathPetId = bigint
@@ -643,6 +673,8 @@ export type GetPetByIdResponse = GetPetByIdStatus200 | GetPetByIdStatus400 | Get
 
 /**
  * @description ID of pet that needs to be updated
+ *
+ * Format: `int64`
  * @type integer
  */
 export type UpdatePetWithFormPathPetId = bigint
@@ -708,6 +740,8 @@ export type DeletePetHeaderApiKey = string | undefined
 
 /**
  * @description Pet id to delete
+ *
+ * Format: `int64`
  * @type integer
  */
 export type DeletePetPathPetId = bigint
@@ -755,6 +789,8 @@ export type DeletePetResponse = DeletePetStatus400
 
 /**
  * @description ID of pet to update
+ *
+ * Format: `int64`
  * @type integer
  */
 export type UploadFilePathPetId = bigint
@@ -771,6 +807,8 @@ export type UploadFileQueryAdditionalMetadata = string | undefined
 export type UploadFileStatus200 = ApiResponse
 
 /**
+ * @description
+ * Format: `binary`
  * @type string | undefined
  */
 export type UploadFileData = Blob | undefined
@@ -954,6 +992,8 @@ export type PlaceOrderPatchResponse = PlaceOrderPatchStatus200 | PlaceOrderPatch
 
 /**
  * @description ID of order that needs to be fetched
+ *
+ * Format: `int64`
  * @type integer
  */
 export type GetOrderByIdPathOrderId = bigint
@@ -1008,6 +1048,8 @@ export type GetOrderByIdResponse = GetOrderByIdStatus200 | GetOrderByIdStatus400
 
 /**
  * @description ID of the order that needs to be deleted
+ *
+ * Format: `int64`
  * @type integer
  */
 export type DeleteOrderPathOrderId = bigint