+++ examples/faker/src/gen/models/User.ts
@@ -8,6 +8,8 @@
  */
 export type User = {
   /**
+   * @description
+   * Format: `int64`
    * @example 10
    * @type integer | undefined
    */
@@ -44,6 +46,8 @@ export type User = {
   phone?: string
   /**
    * @description User Status
+   *
+   * Format: `int32`
    * @example 1
    * @type integer | undefined
    */