+++ examples/mcp/src/gen/models/ts/User.ts
@@ -8,6 +8,8 @@
  */
 export type User = {
   /**
+   * @description
+   * Format: `int64`
    * @example 10
    * @type integer | undefined
    */
@@ -18,6 +20,8 @@ export type User = {
    */
   username?: string
   /**
+   * @description
+   * Format: `uuid`
    * @deprecated
    * @type string | undefined
    */
@@ -33,6 +37,8 @@ export type User = {
    */
   lastName?: string
   /**
+   * @description
+   * Format: `email`
    * @example john@email.com
    * @type string | undefined
    */
@@ -49,6 +55,8 @@ export type User = {
   phone?: string
   /**
    * @description User Status
+   *
+   * Format: `int32`
    * @example 1
    * @type integer | undefined
    */