+++ examples/react-query/src/gen/models/ApiResponse.ts
@@ -5,6 +5,8 @@
  */
 export type ApiResponse = {
   /**
+   * @description
+   * Format: `int32`
    * @type integer | undefined
    */
   code?: number