+++ examples/typescript/src/gen6/ts/models/Customer.ts
@@ -18,6 +18,8 @@ export type CustomerParamsStatusEnumKey = (typeof customerParamsStatusEnum)[keyo
  */
 export type Customer = {
   /**
+   * @description
+   * Format: `int64`
    * @example 100000
    * @type integer | undefined
    */