+++ tests/3.0.x/__snapshots__/main/resolverCustomPrefix/types/GetItem.ts
@@ -1,6 +1,8 @@
 // Custom banner
 
 /**
+ * @description
+ * Format: `int64`
  * @type integer
 */
 export type CustomGetItemPathId = bigint;