+++ tests/3.0.x/__snapshots__/pluginClient/clientTypeClass/types/Order.ts
@@ -16,21 +16,29 @@ export type OrderStatusEnumKey = (typeof orderStatusEnum)[keyof typeof orderStat
 */
 export type Order = {
     /**
+     * @description
+     * Format: `int64`
      * @example 10
      * @type integer | undefined
     */
     id?: bigint;
     /**
+     * @description
+     * Format: `int64`
      * @example 198772
      * @type integer | undefined
     */
     petId?: bigint;
     /**
+     * @description
+     * Format: `int32`
      * @example 7
      * @type integer | undefined
     */
     quantity?: number;
     /**
+     * @description
+     * Format: `date-time`
      * @type string | undefined
     */
     shipDate?: string;