+++ tests/3.0.x/__snapshots__/pluginMcp/petStore/types/Tag.ts
@@ -8,6 +8,8 @@
 */
 export type Tag = {
     /**
+     * @description
+     * Format: `int64`
      * @type integer | undefined
     */
     id?: bigint;